Staging: Add 'Settings' option, AngularJS, extract Ellenoir Application.
This is a commit staging which introduces several additions. A Twitter Bootstrap-
powered navigation bar is added with a 'Settings' option. It doesn't do
much at the moment, apart from presenting a single option to change the
name of the computer.

AngularJS from Gogole CDN is also added which will power the
interactions with the system (e.g., forms).

The controllers for AngularJS are not implemented yet.